Make float dtype explicit in P.1853 oxygen attenuation series
`Ao = ho * go * np.ones_like(Ar)` is float64 today only because
`ho * go` binds first as a float scalar, so the product promotes
regardless of `Ar`'s dtype. That is incidental, not stated: reordering
the operands or a future int-valued `Ar` would silently make `Ao`
integral.

Pin the accumulator to float64 explicitly, matching the fix applied to
`_ITU835_5.standard_pressure` in this branch. This is a no-op today:
seeded runs of `total_attenuation_synthesis` are bit-for-bit identical
before and after, for both integer and float arguments.

+++ itur/models/itu1853.py
@@ -396,7 +396,7 @@ class _ITU1853_1:
         e = Tm * rho / 216.7
         go = gamma0_exact(f, P, rho, Tm).value
         ho, _ = slant_inclined_path_equivalent_height(f, P + e, rho).value
-        Ao = ho * go * np.ones_like(Ar)
+        Ao = ho * go * np.ones_like(Ar, dtype=float)
 
         # Step C15: Synthesize unit variance scintillation time series
         sci_0 = self.scintillation_attenuation_synthesis(Ns * Ts, Ts=1)
